moved all payload type logic from `helpers` and `apps` to its own `types/payload.ts`
