+++ src/app-support/types/helpers.ts
@@ -1,6 +1,3 @@
-import { PlainDescriptor, StorageDescriptor } from "polkadot-api";
-import { ChainId } from "./known-chains";
-
 /**
  * Utility type to expand a type into its properties.
  *
@@ -19,45 +16,14 @@ export type Split<
     : [S];
 
 /**
- *  Union of `[] | [key1] | [key1, key2] | ...` across some set of keys
+ *  Union of `[] | [arg1] | [arg1, arg2] | ...` across some `Args`
  */
-export type PartialArgs<T extends readonly any[]> = T extends [
+export type PartialArgs<Args extends readonly any[]> = Args extends [
     ...infer Rest,
     any
 ]
-    ? PartialArgs<Rest> | T
-    : T;
-
-/**
- * The structure of a payload under `Observables.event`
- */
-type EventPayload<Value> = Value;
-
-/**
- * The structure of a payload under `Observables.storage`
- */
-type StoragePayload<Key, Value> = Expand<{
-    key: Key;
-    value: Value;
-}>;
-
-/**
- * Extract specific type inside of a descriptor tree, given by a list of type strings
- */
-export type PayloadLookup<TreeNode, Path extends readonly string[]> =
-    // Pop top key from path
-    Path extends [infer K, ...infer Rest extends string[]]
-        ? // More path— dig deeper
-          K extends keyof TreeNode
-            ? PayloadLookup<TreeNode[K], Rest>
-            : never
-        : // Storage leaf
-        TreeNode extends StorageDescriptor<infer Key, infer Value, any, any>
-        ? StoragePayload<Key, Value>
-        : // Event leaf
-        TreeNode extends PlainDescriptor<infer Value>
-        ? EventPayload<Value>
-        : never;
+    ? PartialArgs<Rest> | Args
+    : Args;
 
 /**
  * Tests covered by sibling files
